portal/requirements.txt
deta==1.1.0
openai==0.25.0
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.27.1
streamlit==1.14.0
